app works, new db, hash done
