--- migration/versions/4a6e59fcc3e1_create_users_table.py
@@ -1,40 +1,0 @@
-"""create users table
-
-Revision ID: 4a6e59fcc3e1
-Revises: 
-Create Date: 2024-10-12 22:04:33.323245
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4a6e59fcc3e1'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=True),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('login', sa.String(), nullable=False),
-    sa.Column('password', sa.String(), nullable=False),
-    sa.Column('salt', sa.String(), nullable=False),
-    sa.Column('createdAt', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('email'),
-    sa.UniqueConstraint('email'),
-    sa.UniqueConstraint('login')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('users')
-    # ### end Alembic commands ###
