--- migration/versions/dde13891733d_create_users_table.py
@@ -1,30 +1,0 @@
-"""create users table
-
-Revision ID: dde13891733d
-Revises: 4a6e59fcc3e1
-Create Date: 2024-10-12 22:06:11.796893
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'dde13891733d'
-down_revision: Union[str, None] = '4a6e59fcc3e1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint(None, 'users', ['email'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'users', type_='unique')
-    # ### end Alembic commands ###
